move core module files to core folder, preparing for other modules
