+++ include/core/binary_data.h
