+++ src/core/mesh.cpp
@@ -7,7 +7,7 @@
   All rights reserved.
 
 =========================================================================*/
-#include "mesh.h"
+#include "core/mesh.h"
 
 //  --------------------------------------------------------------------------------------
 //  NODE
