+++ src/core/result.cpp
@@ -7,7 +7,7 @@
   All rights reserved.
 
 =========================================================================*/
-#include "result.h"
+#include "core/result.h"
 
 //  --------------------------------------------------------------------------------------
 //  RESULT
